Integrate a bufnet local service
