+++ unit_test/conn_test.go
@@ -19,176 +19,12 @@ package unit_test
 import (
 	"context"
 	"fmt"
-	"net"
 	"testing"
 
-	"github.com/dgraph-io/dgo/v240"
 	"github.com/dgraph-io/dgo/v240/protos/api"
-	"github.com/hypermodeinc/dgraph/v24/x"
 	"github.com/hypermodeinc/modusgraph"
 	"github.com/stretchr/testify/require"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/test/bufconn"
 )
-
-// bufSize is the size of the buffer for the bufconn connection
-const bufSize = 1024 * 1024
-
-// serverWrapper wraps the edgraph.Server to provide proper context setup
-type serverWrapper struct {
-	api.DgraphServer
-	engine *modusgraph.Engine
-}
-
-// Query implements the Dgraph Query method by delegating to the Engine
-func (s *serverWrapper) Query(ctx context.Context, req *api.Request) (*api.Response, error) {
-	// Try to extract namespace from context (using x package from dgraph)
-	nsID, err := x.ExtractNamespace(ctx)
-	var ns *modusgraph.Namespace
-
-	if err != nil || nsID == 0 {
-		// If no namespace in context, use default
-		ns = s.engine.GetDefaultNamespace()
-		fmt.Println("Using default namespace:", ns.ID())
-	} else {
-		// Use the namespace from the context
-		ns, err = s.engine.GetNamespace(nsID)
-		if err != nil {
-			return nil, fmt.Errorf("error getting namespace %d: %w", nsID, err)
-		}
-		fmt.Println("Using namespace from context:", ns.ID())
-	}
-
-	// Check if this is a mutation request
-	if len(req.Mutations) > 0 {
-		fmt.Println("Received mutation request with", len(req.Mutations), "mutations")
-
-		// Process mutations
-		mu := req.Mutations[0] // For simplicity, handle first mutation
-		fmt.Printf("Mutation details - SetNquads: %d bytes, DelNquads: %d bytes, CommitNow: %v\n",
-			len(mu.SetNquads), len(mu.DelNquads), mu.CommitNow)
-
-		if len(mu.SetNquads) > 0 {
-			fmt.Println("SetNquads:", string(mu.SetNquads))
-		}
-
-		// Delegate to the Engine's Mutate method
-		uids, err := ns.Mutate(ctx, req.Mutations)
-		if err != nil {
-			return nil, fmt.Errorf("engine mutation error: %w", err)
-		}
-
-		// Convert map[string]uint64 to map[string]string for the response
-		uidMap := make(map[string]string)
-		for k, v := range uids {
-			uidMap[k] = fmt.Sprintf("%d", v)
-		}
-
-		return &api.Response{
-			Uids: uidMap,
-		}, nil
-	}
-
-	// This is a regular query
-	fmt.Println("Processing query:", req.Query)
-	return ns.Query(ctx, req.Query)
-}
-
-// CommitOrAbort implements the Dgraph CommitOrAbort method
-func (s *serverWrapper) CommitOrAbort(ctx context.Context, tc *api.TxnContext) (*api.TxnContext, error) {
-	// Extract namespace from context
-	nsID, err := x.ExtractNamespace(ctx)
-	var ns *modusgraph.Namespace
-
-	if err != nil || nsID == 0 {
-		// If no namespace in context, use default
-		ns = s.engine.GetDefaultNamespace()
-		fmt.Println("CommitOrAbort using default namespace:", ns.ID())
-	} else {
-		// Use the namespace from the context
-		ns, err = s.engine.GetNamespace(nsID)
-		if err != nil {
-			return nil, fmt.Errorf("error getting namespace %d: %w", nsID, err)
-		}
-		fmt.Println("CommitOrAbort using namespace from context:", ns.ID())
-	}
-
-	fmt.Printf("CommitOrAbort called with transaction: %+v\n", tc)
-
-	// Check if the transaction context is being aborted
-	if tc.Aborted {
-		fmt.Println("Transaction aborted")
-		return tc, nil // Just return as is for aborted transactions
-	}
-
-	// For commit, we need to make a dummy mutation that has no effect but will trigger the commit
-	// This approach uses an empty mutation with CommitNow:true to leverage the Engine's existing
-	// transaction commit mechanism
-	emptyMutation := &api.Mutation{
-		CommitNow: true,
-	}
-
-	// We can't directly attach the transaction ID to the context in this way,
-	// but the Server implementation should handle the transaction context
-	// using the StartTs value in the empty mutation
-
-	// Send the mutation through the Engine
-	_, err = ns.Mutate(ctx, []*api.Mutation{emptyMutation})
-	if err != nil {
-		return nil, fmt.Errorf("error committing transaction: %w", err)
-	}
-
-	fmt.Println("Transaction committed successfully")
-
-	// Return a successful response
-	response := &api.TxnContext{
-		StartTs:  tc.StartTs,
-		CommitTs: tc.StartTs + 1, // We don't know the actual commit timestamp, but this works for testing
-	}
-
-	return response, nil
-}
-
-// setupBufconnServer creates a bufconn listener and starts a gRPC server with the Dgraph service
-func setupBufconnServer(t *testing.T, engine *modusgraph.Engine) (*bufconn.Listener, *grpc.Server) {
-	lis := bufconn.Listen(bufSize)
-	server := grpc.NewServer()
-
-	// Register our server wrapper that properly handles context and routing
-	dgraphServer := &serverWrapper{engine: engine}
-	api.RegisterDgraphServer(server, dgraphServer)
-
-	// Start the server in a goroutine
-	go func() {
-		if err := server.Serve(lis); err != nil {
-			t.Logf("Server exited with error: %v", err)
-		}
-	}()
-
-	return lis, server
-}
-
-// bufDialer is the dialer function for bufconn
-func bufDialer(listener *bufconn.Listener) func(context.Context, string) (net.Conn, error) {
-	return func(ctx context.Context, url string) (net.Conn, error) {
-		return listener.Dial()
-	}
-}
-
-// createDgraphClient creates a Dgraph client that connects to the bufconn server
-func createDgraphClient(ctx context.Context, listener *bufconn.Listener) (*dgo.Dgraph, error) {
-	// Create a gRPC connection using the bufconn dialer
-	conn, err := grpc.DialContext(ctx, "bufnet",
-		grpc.WithContextDialer(bufDialer(listener)),
-		grpc.WithInsecure())
-	if err != nil {
-		return nil, err
-	}
-
-	// Create a Dgraph client
-	dgraphClient := api.NewDgraphClient(conn)
-	return dgo.NewDgraphClient(dgraphClient), nil
-}
 
 // TestDgraphWithBufconn tests Dgraph operations using bufconn
 func TestDgraphWithBufconn(t *testing.T) {
@@ -197,16 +33,11 @@ func TestDgraphWithBufconn(t *testing.T) {
 	require.NoError(t, err)
 	defer engine.Close()
 
-	// Set up the bufconn server with access to the engine
-	listener, server := setupBufconnServer(t, engine)
-	defer server.Stop()
-
-	// Create a context
-	ctx := context.Background()
-
-	// Create a Dgraph client
-	client, err := createDgraphClient(ctx, listener)
+	client, err := engine.GetClient()
 	require.NoError(t, err)
+	defer client.Close()
+
+	ctx := context.Background()
 
 	// Test a simple operation
 	txn := client.NewReadOnlyTxn()
@@ -235,6 +66,11 @@ func TestDgraphWithBufconn(t *testing.T) {
 	require.NoError(t, err)
 	fmt.Println("query after mutation:", resp)
 	_ = txn.Discard(ctx)
+
+	err = client.Alter(context.Background(), &api.Operation{DropAll: true})
+	if err != nil {
+		t.Error(err)
+	}
 }
 
 // TestMultipleDgraphClients tests multiple clients connecting to the same bufconn server
@@ -244,19 +80,17 @@ func TestMultipleDgraphClients(t *testing.T) {
 	require.NoError(t, err)
 	defer engine.Close()
 
-	// Set up the bufconn server with access to the engine
-	listener, server := setupBufconnServer(t, engine)
-	defer server.Stop()
-
 	// Create a context
 	ctx := context.Background()
 
 	// Create multiple clients
-	client1, err := createDgraphClient(ctx, listener)
+	client1, err := engine.GetClient()
 	require.NoError(t, err)
+	defer client1.Close()
 
-	client2, err := createDgraphClient(ctx, listener)
+	client2, err := engine.GetClient()
 	require.NoError(t, err)
+	defer client2.Close()
 
 	// Test that both clients can execute operations
 	txn1 := client1.NewTxn()
